update: refactor AuthService to use consolidated JwtService for token management
