+++ src/main/java/team/themoment/imi/domain/auth/service/AuthService.java
@@ -9,8 +9,7 @@ import team.themoment.imi.domain.auth.exception.SignInFailedException;
 import team.themoment.imi.domain.user.entity.User;
 import team.themoment.imi.domain.user.repository.UserJpaRepository;
 import team.themoment.imi.global.security.jwt.data.TokenDto;
-import team.themoment.imi.global.security.jwt.service.JwtIssueService;
-import team.themoment.imi.global.security.jwt.service.JwtParserService;
+import team.themoment.imi.global.security.jwt.service.JwtService;
 
 import java.util.Date;
 
@@ -18,8 +17,7 @@ import java.util.Date;
 @RequiredArgsConstructor
 public class AuthService {
 
-    private final JwtIssueService jwtIssueService;
-    private final JwtParserService jwtParserService;
+    private final JwtService jwtService;
     private final UserJpaRepository userJpaRepository;
     private final PasswordEncoder passwordEncoder;
 
@@ -27,8 +25,8 @@ public class AuthService {
         User user = userJpaRepository.findByEmail(email)
                 .orElseThrow(SignInFailedException::new);
         if (passwordEncoder.matches(password, user.getPassword())) {
-            TokenDto accessToken = jwtIssueService.issueAccessToken(user.getEmail());
-            String refreshToken = jwtIssueService.issueRefreshToken(user.getEmail());
+            TokenDto accessToken = jwtService.issueAccessToken(user.getEmail());
+            String refreshToken = jwtService.issueRefreshToken(user.getEmail());
             return new LoginResDto(accessToken.token(), refreshToken, accessToken.expiresIn(), new Date().getTime());
         } else {
             throw new SignInFailedException();
@@ -36,11 +34,11 @@ public class AuthService {
     }
 
     public LoginResDto refresh(String refreshToken) {
-        if (jwtParserService.validateRefreshToken(refreshToken)) {
-            String userId = jwtParserService.extractUserId(refreshToken);
-            TokenDto accessToken = jwtIssueService.issueAccessToken(userId);
-            String newRefreshToken = jwtIssueService.issueRefreshToken(userId);
-            jwtParserService.deleteRefreshToken(refreshToken);
+        if (jwtService.validateRefreshToken(refreshToken)) {
+            String userId = jwtService.extractUserId(refreshToken);
+            TokenDto accessToken = jwtService.issueAccessToken(userId);
+            String newRefreshToken = jwtService.issueRefreshToken(userId);
+            jwtService.deleteRefreshToken(refreshToken);
             return new LoginResDto(accessToken.token(), newRefreshToken, accessToken.expiresIn(), System.currentTimeMillis());
         } else {
             throw new InvalidRefreshTokenException();
