--- src/main/java/team/themoment/imi/global/security/jwt/service/JwtIssueService.java
@@ -1,9 +1,0 @@
-package team.themoment.imi.global.security.jwt.service;
-
-import team.themoment.imi.global.security.jwt.data.TokenDto;
-
-public interface JwtIssueService {
-    TokenDto issueAccessToken(String userId);
-
-    String issueRefreshToken(String userId);
-}
