--- src/main/java/team/themoment/imi/global/security/jwt/service/JwtParserService.java
@@ -1,11 +1,0 @@
-package team.themoment.imi.global.security.jwt.service;
-
-public interface JwtParserService {
-    String extractUserId(String token);
-
-    Boolean validateToken(String token);
-
-    Boolean validateRefreshToken(String token);
-
-    void deleteRefreshToken(String token);
-}
