+++ src/main/java/team/themoment/imi/global/security/jwt/service/JwtService.java
@@ -1,37 +1,37 @@
-package team.themoment.imi.global.security.jwt.service.impl;
+package team.themoment.imi.global.security.jwt.service;
 
+import io.jsonwebtoken.Claims;
+import io.jsonwebtoken.ExpiredJwtException;
+import io.jsonwebtoken.JwtException;
 import io.jsonwebtoken.Jwts;
 import io.jsonwebtoken.security.Keys;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
+import team.themoment.imi.global.security.exception.ExpiredAccessTokenException;
+import team.themoment.imi.global.security.exception.InvalidAccessTokenException;
 import team.themoment.imi.global.security.jwt.data.TokenDto;
 import team.themoment.imi.global.security.jwt.entity.RefreshTokenRedisEntity;
 import team.themoment.imi.global.security.jwt.repository.RefreshTokenRedisRepository;
-import team.themoment.imi.global.security.jwt.service.JwtIssueService;
 
 import javax.crypto.SecretKey;
 import java.util.Date;
 import java.util.UUID;
 
 @Service
-public class JwtIssueServiceImpl implements JwtIssueService {
+public class JwtService {
 
     private final RefreshTokenRedisRepository refreshTokenRedisRepository;
-
     private final SecretKey key;
-
     @Value("${spring.security.jwt.access-token-expiration}")
     private Long accessTokenExpiration;
-
     @Value("${spring.security.jwt.refresh-token-expiration}")
     private Long refreshTokenExpiration;
 
-    public JwtIssueServiceImpl(RefreshTokenRedisRepository refreshTokenRedisRepository, @Value("${spring.security.jwt.secret}") String secretKey) {
+    public JwtService(RefreshTokenRedisRepository refreshTokenRedisRepository, @Value("${spring.security.jwt.secret}") String secretKey) {
         this.refreshTokenRedisRepository = refreshTokenRedisRepository;
         this.key = Keys.hmacShaKeyFor(secretKey.getBytes());
     }
 
-    @Override
     public TokenDto issueAccessToken(String userId) {
         Date expirationDate = new Date(System.currentTimeMillis() + accessTokenExpiration);
         String token = Jwts.builder()
@@ -44,7 +44,6 @@ public class JwtIssueServiceImpl implements JwtIssueService {
         return new TokenDto(token, expirationDate.getTime());
     }
 
-    @Override
     public String issueRefreshToken(String userId) {
         Date expirationDate = new Date(System.currentTimeMillis() + refreshTokenExpiration);
         String token = Jwts.builder()
@@ -61,5 +60,32 @@ public class JwtIssueServiceImpl implements JwtIssueService {
                 .build()
         );
         return token;
+    }
+
+    public String extractUserId(String token) {
+        return parseClaims(token).getSubject();
+    }
+
+    public Boolean validateToken(String token) {
+        try {
+            parseClaims(token);
+            return true;
+        } catch (ExpiredJwtException e) {
+            throw new ExpiredAccessTokenException();
+        } catch (JwtException e) {
+            throw new InvalidAccessTokenException();
+        }
+    }
+
+    public Boolean validateRefreshToken(String token) {
+        return refreshTokenRedisRepository.existsById(token);
+    }
+
+    public void deleteRefreshToken(String token) {
+        refreshTokenRedisRepository.deleteById(token);
+    }
+
+    private Claims parseClaims(String token) {
+        return Jwts.parser().verifyWith(key).clock(Date::new).build().parseSignedClaims(token).getPayload();
     }
 }
